refactor : Board Service 추상화
